メモリ不足対策
* score_t を double から float に
* スコア計算用配列を malloc で確保

+++ src/match.h
@@ -3,7 +3,7 @@
 
 #include <math.h>
 
-typedef double score_t;
+typedef float score_t;
 #define SCORE_MAX INFINITY
 #define SCORE_MIN -INFINITY
